test(dedupe): consolidate coinbase order_service tests
Cluster 829ec182749c: Consolidate 10 OrderService test modules into 5
focused modules (net -5 files), all under the <=240 test-file hygiene cap.

New structure:
- test_order_service_place_and_cancel.py (place_order + cancel_order)
- test_order_service_queries.py (get_order + list_orders)
- test_order_service_fills.py (list_fills)
- test_order_service_close_position.py (close_position)
- test_contract_suite_order_service_queries_fills_and_close_position.py

Removed:
- test_orders_place_order.py
- test_orders_cancel_order.py
- test_orders_get_order.py
- test_orders_list_orders.py
- test_orders_list_fills.py
- test_orders_close_position.py
- test_orders_edge_cases.py
- test_order_service_edges.py
- test_contract_suite_order_service_close_position.py
- test_contract_suite_order_service_queries_and_fills.py
