+++ var/agents/testing/index.json
@@ -7,8 +7,8 @@
     "source_test_map": "source_test_map.json"
   },
   "summary": {
-    "total_tests": 6334,
-    "total_files": 980,
+    "total_tests": 6325,
+    "total_files": 975,
     "markers_used": 15
   },
   "quick_commands": {
