+++ var/agents/testing/markers.json
@@ -95,12 +95,12 @@
     ]
   },
   "counts": {
-    "unit": 6169,
+    "unit": 6160,
     "asyncio": 371,
     "property": 82,
     "integration": 79,
     "endpoints": 78,
-    "parametrize": 64,
+    "parametrize": 68,
     "skipif": 8,
     "slow": 4,
     "anyio": 3,
